extract some common code into functions
