+++ src/moonlink_connectors/src/pg_replicate/tests/test_utils.rs
@@ -1,6 +1,8 @@
+use crate::pg_replicate::table::{TableName, TableSchema};
 use crate::pg_replicate::ReplicationClient;
 use std::sync::Arc;
 use tokio::sync::mpsc;
+use tokio_postgres::types::PgLsn;
 use tokio_postgres::{connect, NoTls};
 
 const DEFAULT_DB_URL: &str = "postgresql://postgres:postgres@postgres:5432/postgres";
@@ -30,6 +32,88 @@ pub async fn create_replication_client() -> ReplicationClient {
         }
     });
     replication_client
+}
+
+/// Fetch the `TableSchema` for a given table name within the `public` schema,
+/// using an optional publication context for column filtering.
+pub async fn fetch_table_schema(publication: &str, table_name_str: &str) -> TableSchema {
+    let url = database_url();
+    let (schema_pg_client, schema_conn) = connect(&url, NoTls).await.unwrap();
+    tokio::spawn(async move {
+        if let Err(e) = schema_conn.await {
+            eprintln!("Schema connection error: {e}");
+        }
+    });
+    let mut schema_client = ReplicationClient::from_client(schema_pg_client);
+    let table_name = TableName {
+        schema: "public".to_string(),
+        name: table_name_str.to_string(),
+    };
+    let src_table_id = schema_client
+        .get_src_table_id(&table_name)
+        .await
+        .unwrap()
+        .expect("missing table id for table {table_name}");
+    schema_client
+        .get_table_schema(src_table_id, table_name, Some(publication))
+        .await
+        .unwrap()
+}
+
+/// Spawns a background SQL executor that can be used to submit arbitrary SQL
+/// in desired order with optional delays between them.
+pub fn spawn_sql_executor(database_url: String) -> mpsc::UnboundedSender<String> {
+    let (tx, mut rx) = mpsc::unbounded_channel::<String>();
+    tokio::spawn(async move {
+        let (bg_client, bg_connection) = connect(&database_url, NoTls).await.unwrap();
+        tokio::spawn(async move {
+            bg_connection.await.unwrap();
+        });
+
+        while let Some(sql) = rx.recv().await {
+            bg_client.simple_query(&sql).await.unwrap();
+        }
+    });
+    tx
+}
+
+/// Helper to set replica identity FULL on a table.
+pub async fn set_replica_identity_full(client: &tokio_postgres::Client, table_name: &str) {
+    client
+        .simple_query(&format!("ALTER TABLE {table_name} REPLICA IDENTITY FULL;"))
+        .await
+        .unwrap();
+}
+
+/// Helper to create a publication for a single table.
+pub async fn create_publication_for_table(
+    client: &tokio_postgres::Client,
+    publication: &str,
+    table_name: &str,
+) {
+    client
+        .simple_query(&format!(
+            "CREATE PUBLICATION {publication} FOR TABLE {table_name};"
+        ))
+        .await
+        .unwrap();
+}
+
+/// Create a replication client and ensure a replication slot exists, returning the
+/// client and the confirmed flush LSN to start streaming from.
+pub async fn create_replication_client_and_slot(slot_name: &str) -> (ReplicationClient, PgLsn) {
+    let mut replication_client = create_replication_client().await;
+    // https://github.com/supabase/etl/blob/4da956c6b9be8476a1dbe87a4d88689e0671b7c1/etl/docs/Replication%20in%20Postgres.md?plain=1#L70
+    replication_client
+        .begin_readonly_transaction()
+        .await
+        .unwrap();
+    let slot_info = replication_client
+        .get_or_create_slot(slot_name)
+        .await
+        .unwrap();
+    replication_client.commit_txn().await.unwrap();
+    (replication_client, slot_info.confirmed_flush_lsn)
 }
 
 async fn drop_publication(client: &tokio_postgres::Client, publication: &str) {
